Separated unit test for go1.10+ and go1.9
